+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to this project will be documented in this file.
 - *(ai)* Countriesnow adapter and gemini ai working together
 - *(ai)* Devto adapter and gemini ai working together
 - *(ai)* Exploitdb adapter and gemini ai working together
+- *(ai)* Hackernews adapter and gemini ai working together
 
 ### ⚙️ Miscellaneous Tasks